Refactor how we ship python packages and run tests
