+++ pyproject.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 description = "GitHub Actions module for generating, augmenting and distributing SBOMs"
 authors = ["sbomify <hello@sbomify.com>"]
 readme = "README.md"
-package-mode = false
+packages = [{include = "sbomify_action"}]
 
 [tool.poetry.dependencies]
 python = "^3.10"
@@ -15,7 +15,19 @@ sentry-sdk = "^2.21.0"
 [tool.poetry.group.dev.dependencies]
 isort = "^5.13.2"
 black = "^24.8.0"
+pytest = "^8.0.0"
+pytest-cov = "^4.1.0"
+pytest-mock = "^3.12.0"
+
+[tool.poetry.scripts]
+sbomify-action = "sbomify_action.cli.main:main"
 
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+python_files = ["test_*.py"]
+addopts = "--cov=sbomify_action --cov-report=term-missing --cov-fail-under=80"
+
