--- sbomify.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-source /etc/profile.d/venv.sh
-
-poetry run python /usr/src/app/entrypoint.py
